requirements.txt
pandas
duckdb
pyyaml
rich
streamlit
